# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Springfield AI Command Center API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /logs:
    get:
      operationId: listLogs
      tags: [logs]
      summary: List daily class logs
      parameters:
        - name: teacherId
          in: query
          schema:
            type: integer
        - name: classId
          in: query
          schema:
            type: integer
        - name: sectionId
          in: query
          schema:
            type: integer
        - name: subjectId
          in: query
          schema:
            type: integer
        - name: date
          in: query
          schema:
            type: string
        - name: verificationStatus
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Logs list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogsListResponse"
    post:
      operationId: createLog
      tags: [logs]
      summary: Create daily class log (draft)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLogRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogItem"
  /logs/{id}:
    get:
      operationId: getLog
      tags: [logs]
      summary: Get log detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateLog
      tags: [logs]
      summary: Update log (before verification)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLogRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogItem"
  /logs/{id}/submit:
    post:
      operationId: submitLog
      tags: [logs]
      summary: Submit log for verification
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogItem"
  /logs/{id}/verify:
    post:
      operationId: verifyLog
      tags: [logs]
      summary: Coordinator verifies a log
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyLogRequest"
      responses:
        "200":
          description: Verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogItem"
  /logs/{id}/reject:
    post:
      operationId: rejectLog
      tags: [logs]
      summary: Coordinator rejects a log
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectLogRequest"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogItem"
  /logs/missing:
    get:
      operationId: getMissingLogs
      tags: [logs]
      summary: Get missing logs for a date
      parameters:
        - name: date
          in: query
          schema:
            type: string
        - name: classId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Missing logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MissingLogsResponse"
  /logs/compliance:
    get:
      operationId: getCompliance
      tags: [logs]
      summary: Teacher compliance report
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
        - name: teacherId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Compliance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplianceResponse"

  # ── Student Log Events ────────────────────────────────────────────────────────
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: Get my notifications
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationsListResponse"
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
      required: [page, limit, total, totalPages]

    # ── Auth ────────────────────────────────────────────────────────────────────
...
    LogItem:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        branchId:
          type: integer
        teacherId:
          type: integer
        teacherName:
          type: string
        classId:
          type: integer
        className:
          type: string
        sectionId:
          type: integer
        sectionName:
          type: string
        subjectId:
          type: integer
        subjectName:
          type: string
        date:
          type: string
        periodNumber:
          type: integer
        syllabusId:
          type: integer
          nullable: true
        topicPlanned:
          type: string
          nullable: true
        topicTaught:
          type: string
          nullable: true
        syllabusStatus:
          type: string
          nullable: true
        teachingMethod:
          type: string
          nullable: true
        homeworkGiven:
          type: boolean
        homeworkDetails:
          type: string
          nullable: true
        notebookWorkGiven:
          type: boolean
        notebookWorkDetails:
          type: string
          nullable: true
        disciplineIssue:
          type: boolean
        disciplineDetails:
          type: string
          nullable: true
        achievementDetails:
          type: string
          nullable: true
        improvementDetails:
          type: string
          nullable: true
        remarks:
          type: string
          nullable: true
        verificationStatus:
          type: string
        verifiedBy:
          type: integer
          nullable: true
        verifierName:
          type: string
          nullable: true
        verificationTime:
          type: string
          format: date-time
          nullable: true
        coordinatorRemarks:
          type: string
          nullable: true
        submittedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, tenantId, branchId, teacherId, teacherName, classId, className, sectionId, sectionName, subjectId, subjectName, date, periodNumber, homeworkGiven, notebookWorkGiven, disciplineIssue, verificationStatus, createdAt, updatedAt]
    LogDetail:
      allOf:
        - $ref: "#/components/schemas/LogItem"
        - type: object
          properties:
            studentEvents:
              type: array
              items:
                $ref: "#/components/schemas/EventItem"
          required: [studentEvents]
    CreateLogRequest:
      type: object
      properties:
        branchId:
          type: integer
        classId:
          type: integer
        sectionId:
          type: integer
        subjectId:
          type: integer
        date:
          type: string
        periodNumber:
          type: integer
        syllabusId:
          type: integer
        topicPlanned:
          type: string
        topicTaught:
          type: string
        syllabusStatus:
          type: string
        teachingMethod:
          type: string
        homeworkGiven:
          type: boolean
        homeworkDetails:
          type: string
        notebookWorkGiven:
          type: boolean
        notebookWorkDetails:
          type: string
        disciplineIssue:
          type: boolean
        disciplineDetails:
          type: string
        achievementDetails:
          type: string
        improvementDetails:
          type: string
        remarks:
          type: string
      required: [branchId, classId, sectionId, subjectId, date, periodNumber]
    UpdateLogRequest:
      type: object
      properties:
        topicPlanned:
          type: string
        topicTaught:
          type: string
        syllabusStatus:
          type: string
        teachingMethod:
          type: string
        homeworkGiven:
          type: boolean
        homeworkDetails:
          type: string
        notebookWorkGiven:
          type: boolean
        notebookWorkDetails:
          type: string
        disciplineIssue:
          type: boolean
        disciplineDetails:
          type: string
        achievementDetails:
          type: string
        improvementDetails:
          type: string
        remarks:
          type: string
    VerifyLogRequest:
      type: object
      properties:
        coordinatorRemarks:
          type: string
    RejectLogRequest:
      type: object
      properties:
        coordinatorRemarks:
          type: string
      required: [coordinatorRemarks]
    LogsListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/LogItem"
        pagination:
          $ref: "#/components/schemas/Pagination"
      required: [data, pagination]
    MissingLogsResponse:
      type: object
      properties:
        date:
          type: string
        missing:
          type: array
          items:
            type: object
            properties:
              teacherId:
                type: integer
              teacherName:
                type: string
              classId:
                type: integer
              className:
                type: string
              sectionId:
                type: integer
              sectionName:
                type: string
              subjectId:
                type: integer
              subjectName:
                type: string
            required: [teacherId, teacherName, classId, className, sectionId, sectionName, subjectId, subjectName]
        totalMissing:
          type: integer
      required: [date, missing, totalMissing]
    ComplianceResponse:
      type: object
      properties:
        dateFrom:
          type: string
        dateTo:
          type: string
        teachers:
          type: array
          items:
            type: object
            properties:
              teacherId:
                type: integer
              teacherName:
                type: string
              expected:
                type: integer
              submitted:
                type: integer
              compliancePercent:
                type: number
            required: [teacherId, teacherName, expected, submitted, compliancePercent]
      required: [dateFrom, dateTo, teachers]

    # ── Events ───────────────────────────────────────────────────────────────────
    EventItem:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        dailyLogId:
          type: integer
        studentId:
          type: integer
        studentName:
          type: string
        admissionNo:
          type: string
        teacherId:
          type: integer
        teacherName:
          type: string
        eventType:
          type: string
        severity:
          type: string
        remarks:
          type: string
          nullable: true
        followUpRequired:
          type: boolean
        followUpDate:
          type: string
          nullable: true
        status:
          type: string
        logDate:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, tenantId, dailyLogId, studentId, studentName, admissionNo, teacherId, teacherName, eventType, severity, followUpRequired, status, createdAt, updatedAt]
...
    NotificationItem:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        userId:
          type: integer
        title:
          type: string
        body:
          type: string
        type:
          type: string
        isRead:
          type: boolean
        relatedEntityType:
          type: string
          nullable: true
        relatedEntityId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, tenantId, userId, title, body, type, isRead, createdAt]
    NotificationsListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/NotificationItem"
        unreadCount:
          type: integer
      required: [data, unreadCount]

    # ── Audit Logs ────────────────────────────────────────────────────────────────
...